requirements.txt
Authlib==0.14.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
Flask
requests
python-dotenv
flask-login
raw
Flask-SQLAlchemy
psycopg2-binary
google-api-python-client==2.30.0
google-auth-httplib2==0.1.0
<<<<<<< HEAD
google-auth-oauthlib==0.4.6
=======
google-auth-oauthlib==0.4.6
arrow

>>>>>>> 174fff7daba68dc4ae55671c6dde4a0331be8f1d
